JBTIS-1026: Extend and use runtime annotation
This adds necessarry classes for Runtime annotation to be usable.
Adds handler for Jbpm Runtime Preference page and its Wizard.
Adds an option to distinquish between MVN and RUNTIME options
in new  project setup wizard and makes use of Runtime annotation
to handle runtime setup.
Adds rd.config system property.
Adds necessary lines for rd.config property to work correctlly.
Needs to be specified as a new maven property and then used as
system property for RedDeer to find it and asociate it with to
rd.config.